gh-65 Move all reference data resources
