+++ src/index.ts
@@ -23,6 +23,7 @@ export * from './mdm-model-types.resource';
 export * from './mdm-container-types.model';
 export * from './data-models/index';
 export * from './terminologies/index';
+export * from './reference-data/index';
 export * from './mdm-merge-types.model';
 export * from './mdm-classifier.model';
 export * from './mdm-classifier.resource';
@@ -36,11 +37,6 @@ export * from './mdm-provider.model';
 export * from './mdm-provider.resource';
 export * from './mdm-catalogue-item.model';
 export * from './mdm-catalogue-item.resource';
-export * from './mdm-reference-data-model.model';
-export * from './mdm-reference-data-model.resource';
-export * from './mdm-reference-data-element.resource';
-export * from './mdm-reference-data-type.resource';
-export * from './mdm-reference-data-value.resource';
 export * from './mdm-version-link.resource';
 export * from './mdm-admin.model';
 export * from './mdm-admin.resource';
