+++ src/mdm-import-export.model.ts
@@ -19,7 +19,7 @@ SPDX-License-Identifier: Apache-2.0
 
 import { CodeSetDetail } from './terminologies/mdm-code-set.model';
 import { DataModelDetail } from './data-models/mdm-data-model.model';
-import { ReferenceDataModelDetail } from './mdm-reference-data-model.model';
+import { ReferenceDataModelDetail } from './reference-data/mdm-reference-data-model.model';
 import { TerminologyDetail } from './terminologies/mdm-terminology.model';
 import {
   AsyncParameters,
