+++ src/reference-data/index.ts
@@ -1,0 +1,5 @@
+export * from './mdm-reference-data-element.resource';
+export * from './mdm-reference-data-model.model';
+export * from './mdm-reference-data-model.resource';
+export * from './mdm-reference-data-type.resource';
+export * from './mdm-reference-data-value.resource';
