+++ src/reference-data/mdm-reference-data-model.resource.ts
@@ -16,16 +16,16 @@ limitations under the License.
 
 SPDX-License-Identifier: Apache-2.0
 */
-import { FinalisePayload, ModelUpdatePayload } from './mdm-model-types.model';
+import { FinalisePayload, ModelUpdatePayload } from '../mdm-model-types.model';
 import {
   QueryParameters,
   RequestSettings,
   SearchQueryParameters,
   Uuid
-} from './mdm-common.model';
-import { MdmModelDomainResource } from './mdm-model-types.resource';
-import { MdmResourcesConfiguration } from './mdm-resources-configuration';
-import { MdmRestHandler } from './mdm-rest-handler';
+} from '../mdm-common.model';
+import { MdmModelDomainResource } from '../mdm-model-types.resource';
+import { MdmResourcesConfiguration } from '../mdm-resources-configuration';
+import { MdmRestHandler } from '../mdm-rest-handler';
 
 /**
  * MDM resource for the management of Reference Data Models.
