+++ src/reference-data/mdm-reference-data-value.resource.ts
@@ -16,8 +16,12 @@ limitations under the License.
 
 SPDX-License-Identifier: Apache-2.0
 */
-import { RequestSettings, QueryParameters, SearchQueryParameters } from './mdm-common.model';
-import { MdmResource } from './mdm-resource';
+import {
+  RequestSettings,
+  QueryParameters,
+  SearchQueryParameters
+} from '../mdm-common.model';
+import { MdmResource } from '../mdm-resource';
 
 /**
  * Controller: referenceDataModel
@@ -28,31 +32,53 @@ import { MdmResource } from './mdm-resource';
  |   DELETE | /api/referenceDataModels/${referenceDataModelId}/referenceDataValues/${referenceDataValueId}                                                | Action: delete
  */
 export class MdmReferenceDataValueResource extends MdmResource {
+  list(
+    referenceDataModelId: string,
+    queryStringParams?: QueryParameters,
+    restHandlerOptions?: RequestSettings
+  ) {
+    const url = `${this.apiEndpoint}/referenceDataModels/${referenceDataModelId}/referenceDataValues`;
+    return this.simpleGet(url, queryStringParams, restHandlerOptions);
+  }
 
-    list(referenceDataModelId: string, queryStringParams?: QueryParameters, restHandlerOptions?: RequestSettings) {
-        const url = `${this.apiEndpoint}/referenceDataModels/${referenceDataModelId}/referenceDataValues`;
-        return this.simpleGet(url, queryStringParams, restHandlerOptions);
-    }
+  get(
+    referenceDataModelId: string,
+    referenceDataValueId: string,
+    queryStringParams?: QueryParameters,
+    restHandlerOptions?: RequestSettings
+  ) {
+    const url = `${this.apiEndpoint}/referenceDataModels/${referenceDataModelId}/referenceDataValues/${referenceDataValueId}`;
+    return this.simpleGet(url, queryStringParams, restHandlerOptions);
+  }
 
-    get(referenceDataModelId: string, referenceDataValueId: string, queryStringParams?: QueryParameters, restHandlerOptions?: RequestSettings) {
-        const url = `${this.apiEndpoint}/referenceDataModels/${referenceDataModelId}/referenceDataValues/${referenceDataValueId}`;
-        return this.simpleGet(url, queryStringParams, restHandlerOptions);
-    }
+  save(
+    referenceDataModelId: string,
+    data: any,
+    restHandlerOptions?: RequestSettings
+  ) {
+    const url = `${this.apiEndpoint}/referenceDataModels/${referenceDataModelId}/referenceDataValues`;
+    return this.simplePost(url, data, restHandlerOptions);
+  }
 
-    save(referenceDataModelId: string, data: any, restHandlerOptions?: RequestSettings) {
-        const url = `${this.apiEndpoint}/referenceDataModels/${referenceDataModelId}/referenceDataValues`;
-        return this.simplePost(url, data, restHandlerOptions);
-    }
+  update(
+    referenceDataModelId: string,
+    referenceDataValueId: string,
+    data: any,
+    restHandlerOptions?: RequestSettings
+  ) {
+    const url = `${this.apiEndpoint}/referenceDataModels/${referenceDataModelId}/referenceDataValues/${referenceDataValueId}`;
+    return this.simplePut(url, data, restHandlerOptions);
+  }
 
-    update(referenceDataModelId: string, referenceDataValueId: string, data: any, restHandlerOptions?: RequestSettings) {
-        const url = `${this.apiEndpoint}/referenceDataModels/${referenceDataModelId}/referenceDataValues/${referenceDataValueId}`;
-        return this.simplePut(url, data, restHandlerOptions);
-    }
-
-    remove(referenceDataModelId: string, referenceDataValueId: string, queryStringParams?: QueryParameters, restHandlerOptions?: RequestSettings) {
-        const url = `${this.apiEndpoint}/referenceDataModels/${referenceDataModelId}/referenceDataValues/${referenceDataValueId}`;
-        return this.simpleDelete(url, queryStringParams, restHandlerOptions);
-    }
+  remove(
+    referenceDataModelId: string,
+    referenceDataValueId: string,
+    queryStringParams?: QueryParameters,
+    restHandlerOptions?: RequestSettings
+  ) {
+    const url = `${this.apiEndpoint}/referenceDataModels/${referenceDataModelId}/referenceDataValues/${referenceDataValueId}`;
+    return this.simpleDelete(url, queryStringParams, restHandlerOptions);
+  }
 
   /**
    * `HTTP POST` - Search within a reference data model for reference data values that match one or more search terms.
